april/requirements.txt
click
Pillow
pytest
rectpack
ruff
